Refactor and change to use yaml file for storing authentication tokens
